+++ spec/inputs/integration/elasticsearch_spec.rb
@@ -4,7 +4,7 @@ require "logstash/plugin"
 require "logstash/inputs/elasticsearch"
 require_relative "../../../spec/es_helper"
 
-describe LogStash::Inputs::Elasticsearch do
+describe LogStash::Inputs::Elasticsearch, :integration => true do
 
   SECURE_INTEGRATION = ENV['SECURE_INTEGRATION'].eql? 'true'
 
@@ -20,10 +20,7 @@ describe LogStash::Inputs::Elasticsearch do
   let(:password) { ENV['ELASTIC_PASSWORD'] || 'abc123' }
   let(:ca_file) { "spec/fixtures/test_certs/ca.crt" }
 
-  let(:es_url) do
-    es_url = ESHelper.get_host_port
-    SECURE_INTEGRATION ? "https://#{es_url}" : "http://#{es_url}"
-  end
+  let(:es_url) { "http#{SECURE_INTEGRATION ? 's' : nil}://#{ESHelper.get_host_port}" }
 
   let(:curl_args) do
     config['user'] ? "-u #{config['user']}:#{config['password']}" : ''
@@ -46,6 +43,8 @@ describe LogStash::Inputs::Elasticsearch do
     ESHelper.curl_and_get_json_response "#{es_url}/_index_template/*", method: 'DELETE', args: curl_args
     # This can fail if there are no indexes, ignore failure.
     ESHelper.curl_and_get_json_response( "#{es_url}/_index/*", method: 'DELETE', args: curl_args) rescue nil
+    ESHelper.curl_and_get_json_response( "#{es_url}/logs", method: 'DELETE', args: curl_args) rescue nil
+    ESHelper.curl_and_get_json_response "#{es_url}/_refresh", method: 'POST', args: curl_args
   end
 
   shared_examples 'an elasticsearch index plugin' do
@@ -56,6 +55,7 @@ describe LogStash::Inputs::Elasticsearch do
     it 'should retrieve json event from elasticsearch' do
       queue = []
       plugin.run(queue)
+      expect(queue.size).to eq(10)
       event = queue.pop
       expect(event).to be_a(LogStash::Event)
       expect(event.get("response")).to eql(404)
@@ -63,10 +63,6 @@ describe LogStash::Inputs::Elasticsearch do
   end
 
   describe 'against an unsecured elasticsearch', integration: true do
-    before(:each) do
-      plugin.register
-    end
-
     it_behaves_like 'an elasticsearch index plugin'
   end
 
@@ -95,13 +91,13 @@ describe LogStash::Inputs::Elasticsearch do
     end
 
     context 'with ca_file' do
-      let(:config) { super().merge('ssl' => true, 'ca_file' => ca_file) }
+      let(:config) { super().merge('ssl_enabled' => true, 'ssl_certificate_authorities' => ca_file) }
       it_behaves_like 'secured_elasticsearch'
     end
 
     context 'with `ca_trusted_fingerprint`' do
       let(:ca_trusted_fingerprint) { File.read("spec/fixtures/test_certs/ca.der.sha256").chomp }
-      let(:config) { super().merge('ssl' => true, 'ca_trusted_fingerprint' => ca_trusted_fingerprint) }
+      let(:config) { super().merge('ssl_enabled' => true, 'ca_trusted_fingerprint' => ca_trusted_fingerprint) }
 
       if Gem::Version.create(LOGSTASH_VERSION) >= Gem::Version.create("8.3.0")
         it_behaves_like 'secured_elasticsearch'
@@ -125,15 +121,21 @@ describe LogStash::Inputs::Elasticsearch do
 
   context 'setting host:port (and ssl)', secure_integration: true do
 
-    let(:client_options) { { :ca_file => ca_file, :user => user, :password => password } }
+    let(:client_options) { { :ssl_certificate_authorities => ca_file, :user => user, :password => password } }
 
     let(:config) do
       config = super().merge "hosts" => [ESHelper.get_host_port]
-      config.merge('user' => user, 'password' => password, 'ssl' => true, 'ca_file' => ca_file)
+      config.merge('user' => user, 'password' => password, 'ssl_enabled' => true, 'ssl_certificate_authorities' => ca_file)
     end
 
     it_behaves_like 'an elasticsearch index plugin'
 
   end
 
+  describe 'slice', integration: true do
+    let(:config) { super().merge('slices' => 2, 'size' => 2) }
+    let(:plugin) { described_class.new(config) }
+
+    it_behaves_like 'an elasticsearch index plugin'
+  end
 end